pyproject.toml
[tool.poetry]
name = "referral-api"
version = "0.1.0"
description = ""
authors = ["temirovazat <azatbekt5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-phonenumber-field = "^7.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
django-split-settings = "1.1"
phonenumbers = "^8.13.18"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
autopep8 = "^2.0.2"
isort = "^5.12.0"
pre-commit = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
